changed database directory
